requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.2
django-filter==25.1
django-import-export==4.3.7
django-simple-history==3.8.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
gunicorn==23.0.0
openpyxl==3.1.5
packaging==24.2
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
whitenoise==6.9.0
